refactor: `create_at`, `update_at`을 관리하는 `BaseTimeEntity` 구현 및 적용
